Fixed conversion loss from float to int

+++ src/functions.cpp
@@ -4,7 +4,7 @@
 // Code from https://stackoverflow.com/questions/14539867/how-to-display-a-progress-indicator-in-pure-c-c-cout-printf
 void progressBar(float percent) {
 	std::cout << "[";
-	int pos = BAR_WIDTH * percent;
+	int pos = (int)(BAR_WIDTH * percent);
 	for (int i = 0; i < BAR_WIDTH; ++i) {
 		if (i <= pos) std::cout << "=";
 		else std::cout << " ";
